+++ src/main/java/io/github/astrapi69/color/model/RgbFloatModel.java
@@ -26,34 +26,24 @@ package io.github.astrapi69.color.model;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 import lombok.experimental.FieldDefaults;
+import lombok.experimental.SuperBuilder;
 
 /**
- * The class {@link RgbModel} is the data model for the representation of the RGB. RGB stands for
- * Red, Green, and Blue, and it's a color model used in various devices and applications
+ * The class {@link RgbFloatModel} extends from {@link GenericRgbModel} and all values are
+ * {@link Float} objects
  */
 @Getter
 @Setter
-@EqualsAndHashCode
 @ToString
 @NoArgsConstructor
-@AllArgsConstructor
-@Builder(toBuilder = true)
+@SuperBuilder
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class RgbModel
+public class RgbFloatModel extends GenericRgbModel<Float>
 {
-	/** the int value for the red color */
-	int red;
-
-	/** the int value for the green color */
-	int green;
-
-	/** the int value for the blue color */
-	int blue;
 }